Initial implementation of rax:captureHeader extension.
We create an extension to WADL (rax:captureHeader) that says copy the
contents of a WADL parameter to a specific header.  This extension can
be placed on any WADL parameter we currently support.

So for example, if I have

<resource path="/path/to/my/{device}">
   <param name="device" style="template"
   rax:captureHeader="X-DEVICE-ID"/>
       ...
</resource>

Then when I make the request:

GET /path/to/my/3778

the request will be extended with the header:

X-DEVICE-ID=3778