Update/Cleanup project layout
This does not include the license or style checkers, but it does break
apart all of the classes into separate files if they are used in more
than one place. Some of the Classes/Objects that are pseudo-private,
were left in place to avoid confusion of there intent to be used
elsewhere.

Some additional enhancements:

1. AffectsChecker is now private to Config
2. Single consolidated ByteArrayServletInputStream instead of having a
separate class for test.
3. Optimized imports on main XSL, XSD, and other source files

This closes #250.