Introduced HEADER_ALL and set to default behavior.
Previously repeating headers were handled with a combination of
HEADER_ANY or HEADERXSD_ANY steps.  These steps ensure that a single
value in the list of all header values must validate correctly.  This
is a useful behavior, and the basis of the rax:roles behavior -- but
it is not correct according to WADL.

We introduce HEADER_ALL which ensures that if there are multiple
checks on a repeating header -- then ALL header values must validate
against those checks.

The only behavior can be brought back by using rax:anyMatch=true on a
header parameter, and we leverage this to get rax:roles to continue to
work.