Initial implementation of rax:assert extension
The extension allows specifying an XPath3.1 assertion.  The assertion
works similar to XSD 1.1 assertion and supports @message, @code for
reporting errors.

The assertion may look at the uri ($req:uri), method ($req:method),
request body ($_ or $body), and headers (with a call to req:header or
req:headers).  The extension may be placed in a
wadl:method/wadl:request,
wadl:method/wadl:request/wadl:representation.  The extension may also
be placed in wadl:resources where it applies to all methods globally
or on wadl:resource which applies to all methods of that resource and
if @applyToChildren is set to true then it will also apply to all
subresources as well.

Note that having an assertion on a wadl:representation will enable
wellformed checks.