Fixed Bug where disconnected checker paths were not causing a fail.
The assert function which asserts that there is a path from one step
to another in a checker file was not asserting in the case where two
separate paths were missing a connection.  I must have been drunk or
tired that day in 2012 when I wrote the assertion code :-)

Luckily this did not leak any actual errors in the code, however some
assert statements in existing tests were incorrect and these have been
fixed.