Added support for rax:captureHeader
This allows setting XPath 3.1 paths at the method request,
representation, resource, and resources level to allow capturing a
header with the result of the XPath. The XPaths can simultaneously
introspect headers, uri, methods, and body content (xml or json).

XPaths work in the exact same way that rax:assert XPaths work, except
that a result sequence is expected which turns into header values.