Extends rax:roles extensions to implement multi-tenant check.
The general pattern is followed:

```xml
    <application xmlns="http://wadl.dev.java.net/2009/02" xmlns:rax="http://docs.rackspace.com/api">
     <resources base="https://test.api.openstack.com">
        <resource path="/v1/{tenant}/resource"
                  rax:roles="a:admin/{tenant}">
            <param name="tenant" style="template"/>
            <method name="POST" rax:roles="a:creator/{tenant}"/>
            <method name="GET" rax:roles="a:observer/{tenant} observer"/>
            <method name="PUT" rax:roles="a:creator/{tenant}"/>
            <method name="DELETE" rax:roles="a:creator/{tenant}"/>
        </resource>
     </resources>
    </application>
```

Note that roles such as `a:creator/{tenant}`. The tenant can be
specified from a parameter of types: template, plain, header or it can
be specified via a the Capture Header extension.

It is still possible to specify a non-tenant role such as `a:creator`,
this signifies (as before) that the role is contained by *any* tenant.

Relevant roles are reported either in the tenanted (`a:creator/{tenant}`) or the
non-tenated (`a:creator`) form to denote whether a tenant check was
performed or not.

If the parameter resolves to more than one tenant value, then
appropriate roles must assigned to *all* the tenants specified or else
the request is rejected.

The feature requires the header `X-MAP-ROLES` to be specified.  The
format is a base64 encoded object of the form:

```javascript
      {
         "1" : ["a:admin","foo","bar"],
         "2" : ["a:creator", "foo", "a:observer"],
         "3" : ["a:updater", "bar", "biz", "a:creator"],
         "4" : ["a:observer"],
         "5" : ["a:admin", "bar", "biz", "a:creator"]
      }
```

..which contains a mapping from tenants to appropriate roles.

The following limitations exists:

1. Tenant role checks are currently not supported when roles are
   masked with 404 or 405 responses. If the mask feature is enabled,
   and multi-tenant roles are specified, the validator will throw an
   error while loading the WADL.

2. Plain parameters can only capture single values in XML and JSON. It
   is, however, possible to specify an XPath into the body with the
   CaptureHeader extension to capture multiple tenants in the XML/JSON
   body.